alloc: fix calculation of reserved pages
The end address of KIP memory descriptors is inclusive. For reserved
regions we want to round up to the next page boundary. In case md.end()
is pointing to the first byte of the page, the calculation was wrong.
Note that this does not make a difference in practice. The lower 10
bits of the end address are always set so this case is never hit.

Change-Id: I53c48dc0f38ea763104040b7bc108ec833cb86cc