vm_task: map load region too
A ram region can have a load offset if the SOC requires the
read-writable access to a memory region be done at a different address
than the read-execute access.

In case the VM needs read-writable access too, it requires access to
the load region. Because the region is already marked as used, it
cannot be added simply by another "ram" command without offset.

Instead, we grant access to the load region unconditionally. The VM
already has access anyway so there is no need to constrain the
alternate access window.

Change-Id: I8f11c089510a7b811dae245cd9ce030190ebeb93